requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
torch==2.1.0
torchvision==0.16.0
pillow==10.2.0
numpy==1.26.4
